+++ Warptex/Vector2D.h
@@ -3,7 +3,7 @@
 class Vector2D {
 public:
 	Vector2D() { Zero(); }
-	Vector2D(int initX, int initY) { X = initX; Y = initY; }
+	Vector2D(float initX, float initY) { X = initX; Y = initY; }
 
 	// Allows +Operation between 2 vectors
 	inline Vector2D operator+(const Vector2D& vector) const {
@@ -17,5 +17,5 @@ public:
 
 	// Reset to 0 the X and Y
 	void Zero() { X = 0; Y = 0; }
-	int X, Y;
+	float X, Y;
 };
